requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
beautifulsoup4==4.14.3
bidict==0.23.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
distro==1.9.0
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
google==3.0.0
google-auth==2.49.1
google-auth-oauthlib==1.3.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
oauthlib==3.3.1
openai==2.29.0
packaging==26.0
pillow==12.1.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
python3-openid==3.2.0
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
simple-websocket==1.1.0
sniffio==1.3.1
social-auth-app-django==5.7.0
social-auth-core==4.8.5
soupsieve==2.8.3
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
whitenoise==6.12.0
wsproto==1.3.2
